UI improvements (#2489)
* [Feature] Implemented improvements on the UI

* [Feature] Show Run metrics on the UI
