+++ frontend/src/pages/Project/Details/Settings/constants.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export const CLI_INFO = {
+    header: <h2>CLI</h2>,
+    body: (
+        <>
+            <p>Some text</p>
+        </>
+    ),
+};
