+++ frontend/src/pages/Project/Details/Settings/styles.module.scss
@@ -7,3 +7,17 @@
 .dangerSectionField {
     width: 300px;
 }
+
+.codeWrapper {
+    position: relative;
+
+    .code {
+        padding: 16px 12px;
+    }
+
+    .copy {
+        position: absolute;
+        top: 10px;
+        right: 8px;
+    }
+}
