+++ frontend/src/pages/Project/Gateways/Table/constants.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export const GATEWAYS_INFO = {
+    header: <h2>Gateways</h2>,
+    body: (
+        <>
+            <p>Some text</p>
+        </>
+    ),
+};
