+++ frontend/src/pages/Runs/Details/Jobs/Details/JobDetails/styles.module.scss
@@ -1,13 +1,11 @@
-.page {
-    height: 100%;
-
-    & [class^="awsui_tabs-content"] {
-        display: none;
-    }
+.details {
+    height: calc(100vh - 272px);
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
 
     & > [class^="awsui_layout"] {
         height: 100%;
-
         & > [class^="awsui_content"] {
             display: flex;
             flex-direction: column;
@@ -15,4 +13,10 @@
             height: 100%;
         }
     }
+}
+
+.logs {
+    flex-grow: 1;
+    min-height: 0;
+    max-height: calc(100vh - 380px);
 }
