+++ frontend/src/pages/Runs/Details/RunDetails/styles.module.scss
@@ -1,0 +1,5 @@
+.logs {
+    flex-grow: 1;
+    min-height: 0;
+    max-height: calc(100vh - 480px);
+}
