+++ frontend/src/pages/Runs/index.ts
@@ -1,4 +1,6 @@
 export { RunList } from './List';
-export { RunDetails } from './Details';
+export { RunDetailsPage } from './Details';
+export { RunDetails } from './Details/RunDetails';
+export { JobMetrics } from './Details/Jobs/Metrics';
 export { Logs } from './Details/Logs';
 export { Artifacts } from './Details/Artifacts';
